Rename QuasiNonsingularRelativeOrbitalElements
